fix services problems

+++ TaskMind.Application/Services/EmployeeService.cs
@@ -1,4 +1,5 @@
-﻿using TaskMind.Application.DTOs.Employee;
+﻿using Microsoft.AspNetCore.Mvc.Rendering;
+using TaskMind.Application.DTOs.Employee;
 using TaskMind.Application.Mappers;
 using TaskMind.Application.Repositories.Interfaces;
 using TaskMind.Application.Services.Interfaces;
@@ -9,10 +10,12 @@ namespace TaskMind.Application.Services
     public class EmployeeService : IEmployeeService
     {
         private readonly IEmployeeRepository _employeeRepo;
+        private readonly ITeamRepository _teamRepo;
 
-        public EmployeeService(IEmployeeRepository employeeRepo)
+        public EmployeeService(IEmployeeRepository employeeRepo, ITeamRepository teamRepo)
         {
             _employeeRepo = employeeRepo;
+            _teamRepo = teamRepo;
         }
 
         public async Task<IEnumerable<Employee>> GetAllEmployeesAsync()
@@ -20,6 +23,34 @@ namespace TaskMind.Application.Services
 
         public async Task<Employee?> GetEmployeeByIdAsync(int id)
             => await _employeeRepo.GetByIdAsync(id);
+
+        public async Task<SelectList> GetTeamsForDropdownAsync(int? selectedTeamId = null)
+        {
+            var teams = await _teamRepo.GetAllAsync();
+            return new SelectList(teams, "Id", "Name", selectedTeamId);
+        }
+
+        public async Task<EmployeeEditModel?> GetEmployeeForEditAsync(int id)
+        {
+            var employee = await _employeeRepo.GetByIdAsync(id);
+            if (employee == null) return null;
+
+            var teams = await GetTeamsForDropdownAsync(employee.TeamId);
+
+            return new EmployeeEditModel
+            {
+                Employee = new UpdateEmployeeDto
+                {
+                    Id = employee.Id,
+                    Name = employee.Name,
+                    Skills = string.Join(", ", employee.Skills),
+                    CurrentWorkload = employee.CurrentWorkload,
+                    TaskCompletionSpeed = employee.TaskCompletionSpeed,
+                    TeamId = employee.TeamId
+                },
+                Teams = teams
+            };
+        }
 
         public async Task CreateEmployeeAsync(CreateEmployeeDto dto)
         {
@@ -39,10 +70,15 @@ namespace TaskMind.Application.Services
 
         public async Task DeleteEmployeeAsync(int id)
         {
+            var employee = await _employeeRepo.GetByIdAsync(id);
+            if (employee == null)
+                throw new KeyNotFoundException("Employee not found");
+
             await _employeeRepo.DeleteAsync(id);
         }
 
-        public bool EmployeeExists(int id) => _employeeRepo.IsExist(id);
+        public async Task<bool> ExistsAsync(int id)
+            => await _employeeRepo.ExistsAsync(id);
     }
 
 }
